Feature #2379 Devtools: Create add-on properties XML without installing
I would like to add a new script to Devtools, `createAddOnXML.groovy`,
that can create the add-on properties XML (`${name}.script.xml`)
directly from the add-on build file, without having to go through the
process of packaging and installing the add-on. The advantage is that
this saves a lot of time when you regularly add and change translations
and allows you to quickly change a translation and test it again.

The original `createAddOnXML.groovy` was a stripped down version of
`installScriptAddOn.groovy`, but the disadvantage is that you have to
maintain the overlapping code at 2 different places. So this feature
request also involves changes of `installScriptAddOn.groovy` or other
changes on the Freeplane side.

https://sourceforge.net/p/freeplane/featurerequests/2379/